+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 # Change Log
 
+## 1.0.2 (xxxx/xx/xx)
+* Add: LogItem.getTypeInitial() method
+* Fix: An wrong return string of LogItem.getString() method
+* Fix: Refactoring the code
+
 ## 1.0.1 (2015/10/21)
-* Fix: Errors of getSerializedLogs method
+* Fix: Errors of getSerializedLogs() method
 
 ## 1.0.0 (2015/10/20)
 * First release